Fixes compilation with Boost 1.51.0.
Ivanovic mention the bug and its report in the Gentoo bug tracker [1].
Ported from 1.10 branch 2012-11-03T15:22:21Z!koraq@xs4all.nl.

[1] https://bugs.gentoo.org/show_bug.cgi?id=440742

+++ src/image.cpp
@@ -349,7 +349,11 @@ size_t hash_value(const locator::value& val) {
 	using boost::hash_value;
 	using boost::hash_combine;
 
-	size_t hash = hash_value(val.type_);
+	/*
+	 * Boost 1.51.0 seems not longer accept an enumerate value in its hash
+	 * function so cast it to a type it does like.
+	 */
+	size_t hash = hash_value(static_cast<unsigned>(val.type_));
 	if (val.type_ == locator::FILE || val.type_ == locator::SUB_FILE) {
 		hash_combine(hash, val.filename_);
 	}
